% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_font.R
\name{download_fonts}
\alias{download_fonts}
\title{Download font dependencies for traditional Chinese Rmd PDF output}
\usage{
download_fonts()
}
\description{
This function is expected to run once after the package is
downloaded. It downloads font files into the \code{rmdtk}'s
R Markdown template folders, so when a template is used
later, it will contained the necessary fonts to work.
}
